+++ NEWS.md
@@ -75,6 +75,8 @@
 
 ### `NDArray`
 
+* A port of Python's `autograd` for `NDArray` (#274)
+
 * `size(x, dims...)` is supported now. (#TBD)
 
   ```julia